Add files via upload
The code implements a news application using Tkinter for the GUI and requests for fetching news data from NewsAPI. It creates a window that displays news headlines, along with their images and descriptions. Users can navigate through the news items using "Previous" and "Next" buttons, and read the full articles by clicking on a "Read More" button. The application utilizes PIL for image handling and webbrowser to open articles in the default browser. The code ensures that the navigation buttons are disabled when reaching the first or last news item. The News App provides a simple and interactive way to stay updated with the latest news.

The News App provides a straightforward way to browse and read top headlines from the NewsAPI. With its user-friendly interface, you can quickly access news articles, stay informed, and customize the application to suit your needs. Enjoy staying updated with the latest news using the News App by Kashif!